+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added a `contentSafeAreaOverlapBehavior` property to `ScrollView` that is used with `ContentInsetAdjustmentBehavior.scrollableAxes`.
+
 ### Removed
 
 ### Changed